Former-commit-id: 8196ab8c6cbd43da7fbb2f9ce9d0d1208400e6d0 [formerly 8bc90893420eb3dcef7d53a45a28d633434b8a8e]
Former-commit-id: 9d486fbd09526da5a4ece97d6bc533e76fa97464

+++ SConstruct
@@ -41,14 +41,24 @@ import subprocess
 import SCons.Script
 import SCons.SConf
 
-
+MACOSX = (platform.system() == 'Darwin')
+WINDOWS = (platform.system() == 'Windows')
+LINUX = (platform.system() == 'Linux')
 
 # URL where we have most of our tgz files for libraries, modules and packages.
 URL_BASE = os.environ['SCIPION_URL_SOFTWARE']
 
-# Define our builders.
-download = Builder(action='wget -nv $SOURCE -c -O $TARGET')
-untar = Builder(action='tar -C $cdir --recursive-unlink -xzf $SOURCE')
+# Define our builders
+if LINUX:
+    download = Builder(action='wget -nv --show-progress -c -O $TARGET $SOURCE')
+    untar = Builder(action='tar -C $cdir --recursive-unlink -xzf $SOURCE')
+elif MACOSX:
+    download = Builder(action='curl -L "$SOURCE" -o "$TARGET"')
+    untar = Builder(action='tar -C $cdir -xzf $SOURCE')
+else:
+    print 'OS not tested yet'
+    Exit(1)
+
 
 # Create the environment the whole build will use.
 env = Environment(ENV=os.environ,
@@ -114,13 +124,13 @@ CheckConfigLib = Builder(action=checkConfigLib)
 
 # Add the path to dynamic libraries so the linker can find them.
 
-if platform.system() == 'Linux':
+if LINUX:
     env.AppendUnique(LIBPATH=os.environ.get('LD_LIBRARY_PATH', ''))
-elif platform.system() == 'Darwin':
-    print "OS not tested yet"
+elif MACOSX:
     env.AppendUnique(LIBPATH=os.environ.get('DYLD_FALLBACK_LIBRARY_PATH', ''))
-elif platform.system() == 'Windows':
+elif WINDOWS:
     print "OS not tested yet"
+    Exit(1)
 else:
     print "Unknown system: %s\nPlease tell the developers." % platform.system()
